+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/tasks/InstrumentedJarTask.kt
@@ -6,10 +6,9 @@ import org.gradle.api.Project
 import org.gradle.api.file.DuplicatesStrategy
 import org.gradle.api.tasks.bundling.Jar
 import org.gradle.kotlin.dsl.named
-import org.gradle.kotlin.dsl.the
+import org.jetbrains.intellij.platform.gradle.Constants.Configurations.Attributes.ArchiveClassifier
 import org.jetbrains.intellij.platform.gradle.Constants.INSTRUMENT_CODE
 import org.jetbrains.intellij.platform.gradle.Constants.Tasks
-import org.jetbrains.intellij.platform.gradle.extensions.IntelliJPlatformExtension
 import org.jetbrains.intellij.platform.gradle.tasks.compaion.JarCompanion
 
 abstract class InstrumentedJarTask : Jar() {
@@ -17,20 +16,15 @@ abstract class InstrumentedJarTask : Jar() {
     companion object : Registrable {
         override fun register(project: Project) =
             project.registerTask<InstrumentedJarTask>(Tasks.INSTRUMENTED_JAR) {
-                val extension = project.the<IntelliJPlatformExtension>()
-                val instrumentCodeEnabled = extension.instrumentCode
-
                 val instrumentCodeTaskProvider = project.tasks.named<InstrumentCodeTask>(INSTRUMENT_CODE)
                 val jarTaskProvider = project.tasks.named<Jar>(Tasks.External.JAR)
 
-                from(instrumentCodeTaskProvider)
-                from(project.zipTree(jarTaskProvider.map { it.archiveFile }))
-
                 duplicatesStrategy = DuplicatesStrategy.EXCLUDE
                 archiveClassifier.convention("instrumented")
-
                 JarCompanion.applyPluginManifest(this)
-                JarCompanion.configureConditionalArtifact(this, jarTaskProvider, instrumentCodeEnabled)
+
+                from(instrumentCodeTaskProvider)
+                from(project.zipTree(jarTaskProvider.map { it.archiveFile }))
             }
     }
 }