+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Introduce `VerifyPluginProjectConfigurationTask.hasModulePlugin` to exclude modules using `org.jetbrains.intellij.platform.module` subplugin from `plugin.xml` checks.
 - Better error handling in dependency helpers when missing values
 - Introduce `GenerateManifestTask` for generating `MANIFEST.MF` file
+- Introduce `ComposedJarTask` to compose and pick the final jar archive
 
 ### Fixed